* master: (186 commits)
  added template for release notes
  bump version to 1.2.0 in preparation for next minor release
  added a template for 1.1.1 release notes
  Bumped version number to 1.1.1
  fixed typos in release notes
  Reorganized references
  updated sprokit directory descriptions
  Updated description of directories
  updated MAP-Tk description again
  updated MAP-Tk descriptions
  break long lines
  corrected minor formatting issues
  hide empty acknowledgements
  Fixed Typos
  Point KWIVER_LIBRARY_DIR to where *.lib files live, no matter the OS
  Correct minor style
  Fix some python tests.
  Fix location of Sprokit Python files on Windows
  Minor style
  Add darknet detector fixes
  ...

+++ arrows/core/close_loops_exhaustive.cxx
@@ -200,7 +200,7 @@ close_loops_exhaustive
   for(vital::frame_id_t f = frame_number - 2; f >= last_frame; f-- )
   {
     auto const& matches = all_matches[f].get();
-    int num_matched = static_cast<int>(matches.size());
+    size_t num_matched = matches.size();
     int num_linked = 0;
     if( num_matched >= d_->match_req )
     {
